The size of cipher text is sent before the ciphertext itself.
The size of cipher text is encrypted by the client, send it to the server and after, when the servr will receive it, the encrypted size of ciphertext will be decrypted by the server.

The previous bug is fixed.

Next work to do: the communication has to be dual, i.e. the client and the server can make the same operations: send encrypted bytes and receive them. Of course they can decrypt them.